+++ src/tests/StoreManager.spec.ts
