+++ src/tests/VolumeApp.spec.ts
@@ -1,0 +1,33 @@
+const expect = require("chai").expect;
+import { PasswordApplication } from "../applications/PasswordApp"
+import { Volume } from "../entities/Volume"
+
+
+// const volumeSrc: Volume = new Volume(
+//     "testingVolume",
+//     "/tmp/",
+//     null,
+//     0
+// );
+
+
+
+// describe("application/VolumeApps", () => {
+//     describe("testing volume app", () => {
+
+//     // it("checking for volumes ", () => {
+
+//     //     let volumeApp =  new PasswordApplication();
+//     //     volumeApp.checkSource(volumeSrc)
+
+//     //   expect(1).to.eql(1); // 
+//     // });
+
+
+//     });
+// });
+
+
+
+
+
